\name{plot.mvsusy}
\alias{plot.mvsusy}
\title{ mvsusy plot method }
\description{
  Generate plot(s) for a \code{mvsusy} object.
}
\usage{
  \method{plot}{mvsusy}(x, \dots)
}
\arguments{
  \item{x}{ A \code{mvsusy} object. }
  \item{\dots}{ Ignored. }
}
\details{
  TODO.
}
\value{
  Returns \code{NULL} invisibly. Generate plot(s) as a side effect.
}
\seealso{\code{\link{mvsusy}}}
\examples{
invisible()
}
\keyword{ data }
